Replace list of service account with compute service account
Service account resource was attached to instances as a list of direct reference.
But the attachment also allowed specifying the scope with every attachemnt.

Replaceing the list of service account resource reference with the list of compute service account complex type [has a service account reference and a scope] solves this